vga: fix a lingering misuse of unsigned long -> uint32_t
I don't know how we missed this but it caused lots of problems.
It gets us just a bit further.

+++ sys/src/libmemdraw/defont.c
@@ -58,7 +58,7 @@ getmemdefont(void)
 		return nil;
 	}
 
-	hdr = p+Dy(r)*i->width*sizeof(unsigned long);
+	hdr = p+Dy(r)*i->width*sizeof(uint32_t);
 	n = atoi(hdr);
 	p = hdr+3*12;
 	fc = malloc(sizeof(Fontchar)*(n+1));
